requirements.txt
ansicolors==1.1.8
pandas==1.1.5
docker==5.0.3
psutil==5.6.6